+++ lib/kitchen/directions/bake_iframes/main.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Kitchen
+  module Directions
+    module BakeIframes
+      def self.v1(outer_element:)
+        V1.new.bake(outer_element: outer_element)
+      end
+    end
+  end
+end